Split out 201 RDS resources
