+++ jetbrains-ultimate/src-201/software/aws/toolkits/jetbrains/datagrip/DatagripCompatabilityUtils.kt
